src/main/java/com/robingong/lc101/Solution.java
package com.robingong.lc101;

import com.robingong.common.TreeNode;

/**
 * 给定一个二叉树，检查它是否是镜像对称的。
 */
public class Solution {
    public boolean isSymmetric(TreeNode root) {
        if (root == null) return true;
        return isMirror(root.left, root.right);
    }


    private boolean isMirror(TreeNode p, TreeNode q) {
        if (p == null && q == null) return true;
        if (p == null || q == null) return false;
        return p.val == q.val && isMirror(p.left, q.right) && isMirror(p.right, q.left);
    }
}
